requirements.txt
django
djangorestframework
drf-nested-routers
django-cors-headers
python-dotenv
celery[redis]
requests
